requirements.txt
# Core Dependencies
Flask>=3.0,<4.0
Flask-WTF>=1.2,<2.0
Flask-SQLAlchemy>=3.1,<4.0
python-dotenv>=1.0,<2.0
Faker>=25.0.0
email-validator
weasyprint